# OpenApi/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api Documentation",
    "description": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Inferred Url"
    }
  ],
  "tags": [
    {
      "name": "basic-error-controller",
      "description": "Basic Error Controller"
    },
    {
      "name": "decorator-controller",
      "description": "Decorator Controller"
    },
    {
      "name": "identity-controller",
      "description": "Identity Controller"
    },
    {
      "name": "root-controller",
      "description": "Root Controller"
    }
  ],
  "paths": {
    "/error": {
      "get": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingPUT",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingPOST",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingDELETE",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "options": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingOPTIONS",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "head": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingHEAD",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "patch": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingPATCH",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "trace": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "errorHtml",
        "operationId": "errorHtmlUsingTRACE",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAndView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/callback": {
      "post": {
        "tags": [
          "decorator-controller"
        ],
        "summary": "validationResultCallback",
        "operationId": "validationResultCallbackUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/starthere": {
      "get": {
        "tags": [
          "decorator-controller"
        ],
        "summary": "getInitiatingQrCode",
        "operationId": "getInitiatingQrCodeUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/token": {
      "post": {
        "tags": [
          "decorator-controller"
        ],
        "summary": "token",
        "operationId": "tokenUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/identity": {
      "get": {
        "tags": [
          "identity-controller"
        ],
        "summary": "WholeDoc",
        "operationId": "WholeDocUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "root-controller"
        ],
        "summary": "GetRoot",
        "operationId": "GetRootUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/fakeConfirmationTokenPayload": {
      "get": {
        "tags": [
          "root-controller"
        ],
        "summary": "showConfirmationTokenPayload",
        "operationId": "showConfirmationTokenPayloadUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationTokenPayload"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/fakeInitiatingQrPayload": {
      "get": {
        "tags": [
          "root-controller"
        ],
        "summary": "showInitiatingQrPayload",
        "operationId": "showInitiatingQrPayloadUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InitiatingQrPayload"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CallbackRequestBody": {
        "title": "CallbackRequestBody",
        "required": [
          "confirmationToken"
        ],
        "type": "object",
        "properties": {
          "confirmationToken": {
            "type": "string"
          }
        }
      },
      "ConfirmationTokenPayload": {
        "title": "ConfirmationTokenPayload",
        "required": [
          "category",
          "exp",
          "iat",
          "jti",
          "personalinfodccextract",
          "poa",
          "pod",
          "result",
          "sub",
          "validTo",
          "validationclock",
          "validfrom"
        ],
        "type": "object",
        "properties": {
          "category": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exp": {
            "type": "integer",
            "format": "int64"
          },
          "iat": {
            "type": "integer",
            "format": "int64"
          },
          "jti": {
            "type": "string"
          },
          "personalinfodccextract": {
            "$ref": "#/components/schemas/DccExtract"
          },
          "poa": {
            "type": "string"
          },
          "pod": {
            "type": "string"
          },
          "result": {
            "type": "string"
          },
          "sub": {
            "type": "string"
          },
          "validTo": {
            "type": "string"
          },
          "validationclock": {
            "type": "string"
          },
          "validfrom": {
            "type": "string"
          }
        }
      },
      "DccExtract": {
        "title": "DccExtract",
        "type": "object",
        "properties": {
          "dob": {
            "type": "string"
          },
          "fnt": {
            "type": "string"
          },
          "gnt": {
            "type": "string"
          }
        }
      },
      "InitiatingQrPayload": {
        "title": "InitiatingQrPayload",
        "required": [
          "consent",
          "privacyUrl",
          "protocol",
          "protocolVersion",
          "serviceIdentity",
          "serviceProvider",
          "subject",
          "token"
        ],
        "type": "object",
        "properties": {
          "consent": {
            "type": "string"
          },
          "privacyUrl": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "protocolVersion": {
            "type": "string"
          },
          "serviceIdentity": {
            "type": "string"
          },
          "serviceProvider": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "ModelAndView": {
        "title": "ModelAndView",
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "model": {
            "type": "object"
          },
          "modelMap": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "reference": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "ALREADY_REPORTED",
              "BAD_GATEWAY",
              "BAD_REQUEST",
              "BANDWIDTH_LIMIT_EXCEEDED",
              "CHECKPOINT",
              "CONFLICT",
              "CONTINUE",
              "CREATED",
              "DESTINATION_LOCKED",
              "EXPECTATION_FAILED",
              "FAILED_DEPENDENCY",
              "FORBIDDEN",
              "FOUND",
              "GATEWAY_TIMEOUT",
              "GONE",
              "HTTP_VERSION_NOT_SUPPORTED",
              "IM_USED",
              "INSUFFICIENT_SPACE_ON_RESOURCE",
              "INSUFFICIENT_STORAGE",
              "INTERNAL_SERVER_ERROR",
              "I_AM_A_TEAPOT",
              "LENGTH_REQUIRED",
              "LOCKED",
              "LOOP_DETECTED",
              "METHOD_FAILURE",
              "METHOD_NOT_ALLOWED",
              "MOVED_PERMANENTLY",
              "MOVED_TEMPORARILY",
              "MULTIPLE_CHOICES",
              "MULTI_STATUS",
              "NETWORK_AUTHENTICATION_REQUIRED",
              "NON_AUTHORITATIVE_INFORMATION",
              "NOT_ACCEPTABLE",
              "NOT_EXTENDED",
              "NOT_FOUND",
              "NOT_IMPLEMENTED",
              "NOT_MODIFIED",
              "NO_CONTENT",
              "OK",
              "PARTIAL_CONTENT",
              "PAYLOAD_TOO_LARGE",
              "PAYMENT_REQUIRED",
              "PERMANENT_REDIRECT",
              "PRECONDITION_FAILED",
              "PRECONDITION_REQUIRED",
              "PROCESSING",
              "PROXY_AUTHENTICATION_REQUIRED",
              "REQUESTED_RANGE_NOT_SATISFIABLE",
              "REQUEST_ENTITY_TOO_LARGE",
              "REQUEST_HEADER_FIELDS_TOO_LARGE",
              "REQUEST_TIMEOUT",
              "REQUEST_URI_TOO_LONG",
              "RESET_CONTENT",
              "SEE_OTHER",
              "SERVICE_UNAVAILABLE",
              "SWITCHING_PROTOCOLS",
              "TEMPORARY_REDIRECT",
              "TOO_EARLY",
              "TOO_MANY_REQUESTS",
              "UNAUTHORIZED",
              "UNAVAILABLE_FOR_LEGAL_REASONS",
              "UNPROCESSABLE_ENTITY",
              "UNSUPPORTED_MEDIA_TYPE",
              "UPGRADE_REQUIRED",
              "URI_TOO_LONG",
              "USE_PROXY",
              "VARIANT_ALSO_NEGOTIATES"
            ]
          },
          "view": {
            "$ref": "#/components/schemas/View"
          },
          "viewName": {
            "type": "string"
          }
        }
      },
      "View": {
        "title": "View",
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string"
          }
        }
      }
    }
  }
}